requirements.txt
openai
anthropic
google-genai
requests